requirements.txt
requests>=2.28.0
pycryptodome>=3.15.0
